Clean up API response. Rename util to common.
