+++ x-pack/plugins/synthetics/server/routes/common.ts
